XRPAtom.AI/requirement.txt
requests==2.32.3
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.4